+++ great_international/panels/great_international.py
@@ -1063,6 +1063,25 @@ class AboutUkRegionListingPagePanels:
                 FieldPanel('contact_cta_link'),
             ]
         ),
+        MultiFieldPanel(
+            heading="Explore more of the Investment Atlas section",
+            classname="collapsible",
+            children=[
+                PageChooserPanel(
+                    'related_page_one', [
+                        'great_international.WhyInvestInTheUKPage',
+                        'great_international.InternationalTopicLandingPage',
+                        'great_international.AboutUkRegionPage']),
+                PageChooserPanel('related_page_two', [
+                        'great_international.WhyInvestInTheUKPage',
+                        'great_international.InternationalTopicLandingPage',
+                        'great_international.AboutUkRegionPage']),
+                PageChooserPanel('related_page_three', [
+                        'great_international.WhyInvestInTheUKPage',
+                        'great_international.InternationalTopicLandingPage',
+                        'great_international.AboutUkRegionPage']),
+            ]
+        ),
         SearchEngineOptimisationPanel()
     ]